feat: добавить регистрацию и авторизацию пользователей, добавить валидацию данных

+++ controllers/cards.js
@@ -1,28 +1,35 @@
 const Cards = require('../models/card')
-const { handleError } = require('../utils/errors')
+const { handleError, FORBIDDEN, StatusCodeError } = require('../utils/errors')
 
-const getCards = (req, res) =>
+const getCards = (req, res, next) =>
   Cards.find({})
     .then((card) => res.send(card))
-    .catch((err) => handleError(err, res))
+    .catch((err) => handleError(err, next))
 
-const createCard = (req, res) => {
+const createCard = (req, res, next) => {
   const { name, link } = req.body
   const owner = req.user._id
   return Cards.create({ name, link, owner })
     .then((card) => res.status(201).send(card))
-    .catch((err) => handleError(err, res))
+    .catch((err) => handleError(err, next))
 }
 
-const deleteCard = (req, res) => {
+const deleteCard = (req, res, next) => {
   const { cardId } = req.params
-  return Cards.findByIdAndRemove(cardId)
+
+  return Cards.findById(cardId)
     .orFail()
-    .then((card) => res.send(card))
-    .catch((err) => handleError(err, res))
+    .then((card) => {
+      if (card.owner.toString() === req.user._id)
+        Cards.findByIdAndRemove(cardId)
+          .orFail()
+          .then((card) => res.send(card))
+      else throw new StatusCodeError(FORBIDDEN)
+    })
+    .catch((err) => handleError(err, next))
 }
 
-const toggleLike = (req, res, isLiked = true) => {
+const toggleLike = (req, res, isLiked = true, next) => {
   const { cardId } = req.params
   return Cards.findByIdAndUpdate(
     cardId,
@@ -33,7 +40,7 @@ const toggleLike = (req, res, isLiked = true) => {
   )
     .orFail()
     .then((card) => res.send(card))
-    .catch((err) => handleError(err, res))
+    .catch((err) => handleError(err, next))
 }
 
 module.exports = {
